Move mediapicker
